Change Data Structures and Refactoring
A lot of stuff going on in this commit. The data structures were not
entirely pleasing me. I chose to alter their format, from a typical
edge/vertex structure to only vertices with neighbors embedded.
I think this structure will be easier to use with the A* algorithm.

Some tests were added, to give a more TDD feel to the project. I hope it
is going to make the code more meaningful and usable from an outer
persective.